+++ src/main.cpp
@@ -48,7 +48,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.10.26"};
+constexpr auto version = std::string_view{"0.10.28"};
 
 struct BootstrapConfigResult final
 {